+++ docs/about/whats_new.rst
@@ -8,6 +8,7 @@ What's New
 v1.8.next
 =========
 
+- Clone ``postgres`` index driver as ``postgis``, and flag as experimental. (:pull:`1248`)
 - Implement a local non-persistent in-memory index driver, with maximal backwards-compatibility
   with default postgres index driver. Doesn't work with CLI interface, as every invocation
   will receive a new, empty index, but useful for testing and small scale proof-of-concept
@@ -27,7 +28,7 @@ v1.8.next
 - Migrate test docker image from `datacube/geobase` to `osgeo/gdal`. (:pull:`1233`)
 - Separate index driver interface definition from default index driver implementation. (:pull:`1226`)
 - Prefer WKT over EPSG when guessing CRS strings. (:pull:`1223`)
-- Updates to documentation. (:pull:`1208`, :pull:`1212`, :pull:`1215`, :pull:`1218`, :pull:`1240`)
+- Updates to documentation. (:pull:`1208`, :pull:`1212`, :pull:`1215`, :pull:`1218`, :pull:`1240`, :pull:`1244`)
 - Tweak to segmented in geometry to suppress Shapely warning. (:pull:`1207`)
 - Improve documentation for Product Discovery, Data Searching and Data Loading docs